head() no longer a function

+++ utils/build.sh
@@ -3,8 +3,13 @@
 sudo service docker start
 export REPO=${REPO:-desktop-k:5000/}
 export TAG=${TAG:-$(git rev-parse --short HEAD)}
-docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
-docker buildx bake --push --set '*.platform=linux/arm/v7' -f docker-compose.yml -f docker-compose.fileserver.yml $@
+#export PLATFORM= #linux/arm/v7'
+if [[ $PLATFORM == linux/arm/v7 ]]
+then
+	docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
+fi
+		
+docker buildx bake --push --set "*.platform=$PlATFORM" -f docker-compose.yml -f docker-compose.fileserver.yml $@
 for service in db backend
 do
     docker tag $REPO$service:$TAG $REPO$service:latest
